Ignore gitfs tests when symlinks not enabled
