+++ tests/unit/modules/test_inspect_collector.py
@@ -19,11 +19,10 @@
 # Import Python Libs
 from __future__ import absolute_import
 import os
-import errno
-import subprocess
 
 # Import Salt Testing Libs
 from tests.support.unit import TestCase, skipIf
+from tests.support.helpers import no_symlinks
 from tests.support.mock import (
     MagicMock,
     patch,
@@ -33,31 +32,6 @@ from tests.support.mock import (
 
 # Import salt libs
 from salt.modules.inspectlib.collector import Inspector
-
-
-HAS_SYMLINKS = None
-
-
-def no_symlinks():
-    '''
-    Check if git is installed and has symlinks enabled in the configuration.
-    '''
-    global HAS_SYMLINKS
-    if HAS_SYMLINKS is not None:
-        return not HAS_SYMLINKS
-    output = ''
-    try:
-        output = subprocess.check_output('git config --get core.symlinks', shell=True)
-    except OSError as exc:
-        if exc.errno != errno.ENOENT:
-            raise
-    except subprocess.CalledProcessError:
-        # git returned non-zero status
-        pass
-    HAS_SYMLINKS = False
-    if output.strip() == 'true':
-        HAS_SYMLINKS = True
-    return not HAS_SYMLINKS
 
 
 @skipIf(NO_MOCK, NO_MOCK_REASON)
